+++ Configuration/TSconfig/GridElements.ts
@@ -2,6 +2,7 @@
 
 <INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/AccordionGridElement.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/ColumnLayoutGridElement.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/ConditionallyGridElement.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/DocumentSectionGridElement.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/PanelGridElement.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/TimelineGridElement.ts">