+++ Configuration/TypoScript/GridElements/constants.ts
@@ -2,6 +2,7 @@
 
 <INCLUDE_TYPOSCRIPT: source="FILE: ./AccordionGridElement/constants.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayoutGridElement/constants.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./ConditionallyGridElement/constants.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./DocumentSectionGridElement/constants.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./PanelGridElement/constants.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./TimelineGridElement/constants.ts">