+++ Configuration/TypoScript/GridElements/setup.ts
@@ -2,6 +2,7 @@
 
 <INCLUDE_TYPOSCRIPT: source="FILE: ./AccordionGridElement/setup.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayoutGridElement/setup.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./ConditionallyGridElement/setup.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./DocumentSectionGridElement/setup.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./PanelGridElement/setup.ts">
 <INCLUDE_TYPOSCRIPT: source="FILE: ./TimelineGridElement/setup.ts">